Change length parameter type for CreateEncapsulationStructure().
The int type generated compiler warnings due to implicit conversions
between size_t and int. All calls to this function use size_t for the
length parameter anyway.

+++ source/src/enet_encap/encap.h
@@ -101,7 +101,7 @@ void EncapsulateListIdentityResponseMessage(
   ENIPMessage *const outgoing_message);
 
 EipInt16 CreateEncapsulationStructure(const EipUint8 *receive_buffer,
-                                      int receive_buffer_length,
+                                      size_t receive_buffer_length,
                                       EncapsulationData *const encapsulation_data);
 
 void SkipEncapsulationHeader(ENIPMessage *const outgoing_message);
